+++ src/components/com_base/controllers/behaviors/votable.php
@@ -1,7 +1,7 @@
 <?php
 
 /**
- * LICENSE: ##LICENSE##.
+ * Votable Behavior.
  *
  * @category   Anahita
  *
@@ -10,22 +10,9 @@
  * @copyright  2008 - 2011 rmdStudio Inc./Peerglobe Technology Inc
  * @license    GNU GPLv3 <http://www.gnu.org/licenses/gpl-3.0.html>
  *
- * @version    SVN: $Id: resource.php 11985 2012-01-12 10:53:20Z asanieyan $
- *
  * @link       http://www.GetAnahita.com
  */
 
-/**
- * Votable Behavior.
- *
- * @category   Anahita
- *
- * @author     Arash Sanieyan <ash@anahitapolis.com>
- * @author     Rastin Mehr <rastin@anahitapolis.com>
- * @license    GNU GPLv3 <http://www.gnu.org/licenses/gpl-3.0.html>
- *
- * @link       http://www.GetAnahita.com
- */
 class ComBaseControllerBehaviorVotable extends AnControllerBehaviorAbstract
 {
     /**