+++ src/components/com_locations/views/location/html/default.php
@@ -57,7 +57,7 @@ if (!empty($scope)) {
 }
 ?>
 
-<?= @infinitescroll($item->tagables->fetchSet(), array(
+<?= @infinitescroll($item->taggables->fetchSet(), array(
   'url' => $url,
   'id' => 'an-geolocatables'
 )) ?>