+++ src/components/com_tags/views/tag/html/taggables.php
@@ -1,3 +1,3 @@
 <? defined('ANAHITA') or die('Restricted access') ?>
 
-<?= @helper('ui.nodes', $item->tagables) ?>
+<?= @helper('ui.nodes', $item->taggables) ?>